language_learner_project/requirements.txt
asgiref==3.8.1
black==25.1.0
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
Django==5.2
django-admin-sortable2==2.2.6
django-environ==0.12.0
django-formtools==2.5.1
django-phonenumber-field==5.2.0
django-rosetta==0.10.2
django-widget-tweaks==1.5.0
environ==1.0
idna==3.10
isort==6.0.1
joblib==1.4.2
mypy_extensions==1.1.0
nltk==3.9.1
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.7
polib==1.2.0
pypng==0.20220715.0
regex==2024.11.6
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.4.0
